Fix typed data serialization
This blobs out data of branded values. We need to do something about
branded values, because brands can be recursive, but Spark won't let
us define a recursive schema. We blob out at the toplevel, because it
seems easier to get working immediately. Future optimizations could
work along the lines of keeping some structure, up until the first
level of branding, or up until there is an actual cycle.

The runtime does not yet deal with this. Unbranding is going to do
more work than just throwing away the type tag, we need to recover
deserialized data from the blob as well.

While doing this we get rid of sdata and stype, and go to JSON and
Scala directly. The idea of sdata and stype was to reflect in Coq
what we think the data and type system of Spark are. This was not
really achieved, because we always had hacks for open record blobs in
there anyway. It was also confusing enough to have a bug for
serialization of "special" records, which were serialized as regular
records with special fields, *inside* a regular record.