Potential replacement for scala_of_data
This prints data as a blob, and relies on the runtime to reshape it
into the expected type. Problem is: we have to have the expected type
for this to work. It does not work with the inferred type of the data
because that might be too general. The empty array would be
`fromBlob(..., "[]").asInstanceOf[Array[???]]`
where we do not know what to put in place of the dots or questionmarks.
The inferred type is Bottom, but that does not help us, of course.