Move GSON dependency to JavaService from individual services
The code of JavaService does not need it but just about every service
does, and it makes sense to hold it in common.  In addition to this
"fetched" dependency there is also a statically committed copy of
gson-2.7.jar in qcert/lib.  This is used by the Java runtime and the
samples.  I'm leaving that alone.