Temporary hacks adding techrule and designerrule languages
The idea is to redo this more correctly in terms of the driver
framework, simply noting at extraction time that the translators from
these languages to CAMP are actually written in native ocaml (which in
turn will call out to Java).  Right now, we pretend in places that these
languages "are" CAMP in order to get things through qcert without
causing it to barf.  The callouts to Java are special-cased relatively
high up in the qcert ocaml code