Improve compilation of contants.
Before, each use of an IMP constant lead to a fresh allocation in the
AssemblyScript runtime. Additionally, strings were entirely encoded in
the AST (as function that allocate and inititalize the corresponding
string in the runtime).

Now, constants are serialized into the linear memory of the compiled
module. On first use, a corresponding value is allocated in the memory
of the AssemblyScript runtime. Repeated use of a constant uses the same
value on the runtime side.