Functorize Wasm backend.
Before, the new Wasm backend was intertwined with the extracted QCert code.
The backend could not be be used with Imp definitions extracted in other
projects.

Now, we define which type definitions we need from the extraction in
`wasm_intf.ml`. We abstract from the exact implementation using a functor
(see `wasm_backend.ml`).

Instantiation of the functor for local use happens in `wasm_ast.ml`.
This way we stay compatible with the previous version without touching
the Coq project itself.